backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.2

# Async
httpx==0.27.2
anyio==4.6.0

# Task Queue
celery[redis]==5.4.0
redis==5.1.1

# Crawling & Ingestion
crawl4ai==0.4.0
playwright==1.48.0
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
llama-index==0.11.20
llama-index-vector-stores-postgres==0.2.3

# AI / Embeddings
openai==1.51.2
anthropic==0.36.0
cohere==5.11.0

# Translation
# sarvam-ai: optional dependency — install manually when available
# pip install sarvam-ai  # Uncomment when package is published on PyPI

# Utilities
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==9.0.0
structlog==24.4.0
rich==13.9.1

# Dev/Test
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
ruff==0.7.0
httpx==0.27.2  # For TestClient
faker==30.3.0
factory-boy==3.3.1
